requirements.txt
torch
numpy
tqdm
fire
einops
icecream
Pillow
matplotlib
pandas
openpyxl
neptune
